+++ custom_components/intex_pool/__init__.py
@@ -26,6 +26,7 @@ from homeassistant.exceptions import (
     ConfigEntryNotReady,
     ServiceValidationError,
 )
+from homeassistant.helpers import config_validation as cv
 from homeassistant.helpers import device_registry as dr
 from homeassistant.helpers import entity_registry as er
 from homeassistant.helpers import issue_registry as ir
@@ -99,6 +100,7 @@ _LOGGER = logging.getLogger(__name__)
 
 URL_BASE = "/intex_pool"
 CARD_FILENAME = "intex-pool-card.js"
+CONFIG_SCHEMA = cv.config_entry_only_config_schema(DOMAIN)
 
 
 async def async_setup(hass: HomeAssistant, config: ConfigType) -> bool: